requirements.txt
anthropic>=0.40.0
requests>=2.31.0
atlassian-python-api>=3.41.0
slack-sdk>=3.27.0
google-api-python-client>=2.150.0
google-auth-httplib2>=0.2.0
google-auth-oauthlib>=1.2.0
python-dotenv>=1.0.0
PyYAML>=6.0.1
feedparser>=6.0.0
